refactor: separating styles into a file of their own
